+++ gateway/src/backend/mod.rs
@@ -1,7 +1,8 @@
 mod doubler;
+pub mod events;
 mod start;
 
-use crate::backend_events::{BackendEvents, KafkaShutdownEvent};
+use crate::backend::events::{BackendEvents, KafkaShutdownEvent};
 use crate::idle_status::KafkaActivityObserved;
 pub use doubler::double_commands;
 pub use start::{start_all, BackendInitOptions};
